fix: rates format in result

+++ src/main/java/com/allo/test/model/dto/LatestRateDto.java
@@ -2,13 +2,14 @@ package com.allo.test.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class LatestRateDto {
 
     private String currency;
-    private Double rate;
-    private Double usdBuyRate;
+    private BigDecimal rate;
+    private BigDecimal usdBuyRate;
 
 }
